+++ packages/ws/src/index.ts
@@ -1,4 +1,3 @@
 export * from './configs';
-export * from './cli';
 export { project } from './project';
 export { compile as buildTranslations } from './lib/i18n';