+++ packages/ws/src/options.ts
@@ -1,6 +1,3 @@
-import { CommanderStatic } from 'commander';
-import { LogLevelDesc } from 'loglevel';
-
 export interface Env {
   key: string;
   value: string;
@@ -9,13 +6,5 @@ export interface Env {
 export interface EnvOptions {
   parent: {
     env: Env[];
-  };
-}
-
-export interface BaseOptions extends CommanderStatic, EnvOptions {
-  parent: {
-    logLevel: LogLevelDesc;
-    env: Env[];
-    ignoreUpdate?: true;
   };
 }