+++ packages/ws/src/typings.d.ts
@@ -2,7 +2,6 @@ declare module 'fork-ts-checker-webpack-plugin';
 declare module '@babel/code-frame';
 declare module 'glob-to-regexp';
 declare module 'eslint';
-declare module 'plur';
 declare module 'babel-jest';
 
 declare module 'parent-dirs' {